refactor: move `mod.ts` into `src` directory
make `/mod.ts` for redirect export of `/src/mod.ts`
